backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pypdf2
python-docx
librosa
soundfile
numpy
python-multipart
    
# Speech to Text
openai-whisper
torch
transformers
    
# Metrics
scikit-learn
pandas
joblib
textblob
    
# Audio
pydub
scipy

# Hiring Platform
aiosqlite
reportlab
httpx

